Revert "Plfm-7762"
