+++ src/main/java/org/sagebionetworks/template/repo/kinesis/firehose/KinesisFirehoseConfigValidator.java
@@ -1,7 +1,6 @@
 package org.sagebionetworks.template.repo.kinesis.firehose;
 
 import org.apache.commons.lang3.StringUtils;
-import org.sagebionetworks.template.repo.glue.GlueTableDescriptor;
 
 public class KinesisFirehoseConfigValidator {
