+++ src/main/java/org/sagebionetworks/template/repo/kinesis/firehose/KinesisFirehoseStreamDescriptor.java
@@ -1,7 +1,5 @@
 package org.sagebionetworks.template.repo.kinesis.firehose;
 
-import org.sagebionetworks.template.repo.glue.GlueTableDescriptor;
-
 import java.util.Objects;
 
 public class KinesisFirehoseStreamDescriptor {
