+++ src/main/resources/templates/datewarehouse/etl-jobs-template.json.vpt
@@ -133,14 +133,12 @@
                 },
                 "TableInput": {
                     "Name": "$etl.tableDescriptor.name",
-                    "Description" : "$etl.tableDescriptor.description",
                     "StorageDescriptor": {
                         "Columns": [
-                            #foreach($column in $etl.tableDescriptor.columns)
+                            #foreach($column in $etl.tableDescriptor.columns.entrySet())
                             {
-                                "Name":"$column.name",
-                                "Type":"$column.type",
-                                "Comment":"$column.comment"
+                                "Name":"${column.key}",
+                                "Type":"${column.value}"
                             }
                             #if( $foreach.hasNext),#end
                             #end
@@ -153,11 +151,10 @@
                         "Location": "s3://${stack}.datawarehouse.sagebase.org/${glueDatabaseName}/${etl.tableDescriptor.name}/",
                     },
                     "PartitionKeys": [
-                        #foreach($partition in $etl.tableDescriptor.partitionKeys)
+                        #foreach($partition in $etl.tableDescriptor.partitionKeys.entrySet())
                         {
-                            "Name":"$partition.name",
-                            "Type":"$partition.type",
-                            "Comment":"$partition.comment"
+                            "Name":"${partition.key}",
+                            "Type":"${partition.value}"
                         }
                         #if( $foreach.hasNext),#end
                         #end
