+++ src/main/resources/templates/repo/kinesis-log-streams.json.vpt
@@ -93,19 +93,12 @@
 			},
 			"TableInput": {
 				"Name": "$stream.tableDescriptor.name.toLowerCase()",
-				#if( $stream.tableDescriptor.description )
-				    "Description" : "$stream.tableDescriptor.description",
-				#end
 				"StorageDescriptor": {
 					"Columns": [
-						#foreach($column in $stream.tableDescriptor.columns)
+						#foreach($column in $stream.tableDescriptor.columns.entrySet())
 						{
-							"Name":"$column.name",
-                            "Type":"$column.type"
-                            #if( $column.comment )
-                                ,
-                                "Comment":"$column.comment"
-                            #end
+							"Name":"${column.key}",
+							"Type":"${column.value}"
 						}
 						#if( $foreach.hasNext),#end
 						#end
@@ -139,14 +132,10 @@
 					#end
 				},
 				"PartitionKeys": [
-					#foreach($partition in $stream.tableDescriptor.partitionKeys)
+					#foreach($partition in $stream.tableDescriptor.partitionKeys.entrySet())
 					{
-						"Name":"$partition.name",
-						"Type":"$partition.type"
-						#if( $partition.comment )
-                            ,
-                            "Comment":"$partition.comment"
-                        #end
+						"Name":"${partition.key}",
+						"Type":"${partition.value}"
 					}
 					#if( $foreach.hasNext),#end
 					#end
