+++ src/test/java/org/sagebionetworks/template/datawarehouse/DataWarehouseBuilderImplTest.java
@@ -2,6 +2,7 @@ package org.sagebionetworks.template.datawarehouse;
 
 import com.amazonaws.services.cloudformation.model.Tag;
 import com.amazonaws.services.s3.AmazonS3;
+import com.google.common.collect.ImmutableMap;
 import org.apache.logging.log4j.Logger;
 import org.apache.velocity.app.VelocityEngine;
 import org.json.JSONObject;
@@ -19,14 +20,12 @@ import org.sagebionetworks.template.LoggerFactory;
 import org.sagebionetworks.template.StackTagsProvider;
 import org.sagebionetworks.template.TemplateGuiceModule;
 import org.sagebionetworks.template.config.Configuration;
-import org.sagebionetworks.template.repo.glue.GlueColumn;
-import org.sagebionetworks.template.repo.glue.GlueTableDescriptor;
+import org.sagebionetworks.template.repo.kinesis.firehose.GlueTableDescriptor;
 import org.sagebionetworks.template.utils.ArtifactDownload;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 import java.util.StringJoiner;
@@ -105,16 +104,11 @@ public class DataWarehouseBuilderImplTest {
 		when(etlJobConfig.getGithubRepo()).thenReturn("repo");
 		when(etlJobConfig.getVersion()).thenReturn("1.0.0");
 		when(etlJobConfig.getExtraScripts()).thenReturn(List.of("utilities/utils.py"));
-
-		GlueColumn column = new GlueColumn();
-		column.setName("someColumn");
-		column.setType("string");
-		column.setComment("This is test column");
+		
 		GlueTableDescriptor table = new GlueTableDescriptor();
 		table.setName("someTableRef");
-		table.setDescription("Test table");
-		table.setColumns(Arrays.asList(column));
-		
+		table.setColumns(ImmutableMap.of("someColumn", "string"));
+
 		List<EtlJobDescriptor> jobs = List.of(
 			new EtlJobDescriptor()
 				.withName("testjob")
@@ -170,8 +164,8 @@ public class DataWarehouseBuilderImplTest {
 
 		JSONObject tableProperty = resources.getJSONObject("someTableRefGlueTable").getJSONObject("Properties");
 		assertEquals("{\"Name\":\"someTableRef"
-				+ "\",\"Description\":\"Test table\",\"StorageDescriptor\":{\"Columns\":[{\"Name\":\"someColumn\","
-				+ "\"Type\":\"string\",\"Comment\":\"This is test column\"}],\"InputFormat\":\"org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat\","
+				+ "\",\"StorageDescriptor\":{\"Columns\":[{\"Name\":\"someColumn\","
+				+ "\"Type\":\"string\"}],\"InputFormat\":\"org.apache.hadoop.hive.ql.io.parquet.MapredParquetInputFormat\","
 				+ "\"SerdeInfo\":{\"SerializationLibrary\":" + "\"org.apache.hadoop.hive.ql.io.parquet.serde.ParquetHiveSerDe\"},"
 				+ "\"Compressed\":true,\"Location\":\"s3://dev.datawarehouse.sagebase.org/synapsewarehouse/someTableRef/\"},\"PartitionKeys\":[],\"TableType\":"
 				+ "\"EXTERNAL_TABLE\"}", tableProperty.getString("TableInput"));
